+++ examples/pyspark_local.py
@@ -6,7 +6,7 @@ Requires:
 
 Or from published packages:
 
-    pip install etlantic==0.49.0 etlantic-pyspark==0.49.0
+    pip install etlantic==0.50.0 etlantic-pyspark==0.50.0
 
 Run with: